requirements.txt
django==3.2.0
djangorestframework

